+++ src/io/output_mesh.hh
@@ -23,16 +23,10 @@
 #include "system/sys_profiler.hh"
 #include "input/accessors.hh"
 
-#include "io/output_data_base.hh"
-
-#include "fields/field_values.hh"
-#include "fields/field_set.hh"
-
 #include "tools/general_iterator.hh"
 
 class Mesh;
-template<int, class Value> class Field;
-template<class T> class MeshData;
+template<class T> class ElementDataCache;
 
 class OutputElement;
 typedef GeneralIterator<OutputElement> OutputElementIterator;
@@ -42,16 +36,32 @@ class OutputMesh;
 class OutputMeshDiscontinuous;
 
 
-/// @brief Base class for Output mesh.
 /**
- * Defines common members for OutputMesh classes.
+ * @brief Base class for Output mesh.
+ *
+ * Defines common members for Output mesh classes:
+ *  - OutputMesh represents output mesh with continuous elements
+ *  - OutputMeshDiscontinuous represents output mesh with discontinuous elements
+ *
+ * Making of output meshes and calling of their initialization methods must be execute in correct order, see example:
+@code
+    // Create or get Mesh object
+    Mesh * my_mesh = ...
+
+    // Construct mesh with continuous elements
+    std::make_shared<OutputMesh> output_mesh = std::make_shared<OutputMesh>(*my_mesh);
+    // Creates the mesh identical to the computational one.
+    output_mesh->create_identical_mesh();
+
+    // Construct mesh with discontinuous elements
+    std::make_shared<OutputMeshDiscontinuous> output_mesh_discont = std::make_shared<OutputMeshDiscontinuous>(*my_mesh);
+    // Creates mesh from the given continuous one.
+    output_mesh_discont->create_mesh(output_mesh);
+@endcode
  */
 class OutputMeshBase : public std::enable_shared_from_this<OutputMeshBase>
 {
 public:
-    DECLARE_EXCEPTION(ExcFieldNotScalar, << "Field '" << FieldCommon::EI_Field::qval
-                                         << "' is not scalar in spacedim 3.");
-    
     /// Shortcut instead of spacedim template. We suppose only spacedim=3 at the moment. 
     static const unsigned int spacedim = 3;
     
@@ -72,18 +82,23 @@ public:
     /// Gives iterator to the LAST element of the output mesh.
     OutputElementIterator end();
     
-    /// Selects the error control field out of output field set according to input record.
-    void select_error_control_field(FieldSet &output_fields);
-    
+    /// Creates refined mesh.
+    virtual void create_refined_mesh()=0;
+
+    /// Returns \p error_control_field_name_
+    inline std::string error_control_field_name() {
+    	return error_control_field_name_;
+    }
+
     /// Vector of element indices in the computational mesh. (Important when refining.)
     std::shared_ptr<std::vector<unsigned int>> orig_element_indices_;
     
     /// Vector of node coordinates. [spacedim x n_nodes]
-    std::shared_ptr<MeshData<double>> nodes_;
+    std::shared_ptr<ElementDataCache<double>> nodes_;
     /// Vector maps the nodes to their coordinates in vector @p nodes_.
-    std::shared_ptr<MeshData<unsigned int>> connectivity_;
+    std::shared_ptr<ElementDataCache<unsigned int>> connectivity_;
     /// Vector of offsets of node indices of elements. Maps elements to their nodes in connectivity_.
-    std::shared_ptr<MeshData<unsigned int>> offsets_;
+    std::shared_ptr<ElementDataCache<unsigned int>> offsets_;
     
     /// Returns number of nodes.
     unsigned int n_nodes();
@@ -100,9 +115,9 @@ protected:
     /// Maximal level of refinement.
     const unsigned int max_level_;
     
-    /// Refinement error control field.
-    Field<3, FieldValue<3>::Scalar> *error_control_field_;
-    
+    /// Refinement error control field name.
+    std::string error_control_field_name_;
+
     /// Friend provides access to vectors for element accessor class.
     friend class OutputElement;
 };
@@ -120,7 +135,7 @@ public:
     void create_identical_mesh();
     
     /// Creates refined mesh.
-    void create_refined_mesh();
+    void create_refined_mesh() override;
     
 protected:
     bool refinement_criterion();
@@ -130,6 +145,7 @@ protected:
 };
 
 
+/// @brief Class represents output mesh with discontinuous elements.
 class OutputMeshDiscontinuous : public OutputMeshBase
 {
 public:
@@ -141,7 +157,7 @@ public:
     void create_mesh(std::shared_ptr<OutputMesh> output_mesh);
     
     /// Creates discontinuous refined mesh.
-    void create_refined_mesh();
+    void create_refined_mesh() override;
     
 protected:
     bool refinement_criterion();