+++ unit_tests/fields/field_elementwise_test.cpp
@@ -21,7 +21,7 @@
 #include "system/sys_profiler.hh"
 
 #include "mesh/mesh.h"
-#include "mesh/msh_gmshreader.h"
+#include "io/msh_gmshreader.h"
 
 
 
@@ -47,6 +47,21 @@ string input = R"INPUT(
        TYPE="FieldElementwise",
        gmsh_file="fields/simplest_cube_data.msh",
        field_name="tensor_fixed"
+   },
+   vtk_scalar={
+       TYPE="FieldElementwise",
+       gmsh_file="fields/vtk_ascii_data.vtu",
+       field_name="scalar_field"
+   }
+   vtk_vector={
+       TYPE="FieldElementwise",
+       gmsh_file="fields/vtk_binary_data.vtu",
+       field_name="vector_field"
+   }
+   vtk_tensor={
+       TYPE="FieldElementwise",
+       gmsh_file="fields/vtk_compressed_data.vtu",
+       field_name="tensor_field"
    }
 }
 )INPUT";
@@ -65,16 +80,16 @@ public:
 
         Profiler::initialize();
         
-        FilePath mesh_file( "mesh/simplest_cube.msh", FilePath::input_file);
-        mesh = mesh_constructor();
-        ifstream in(string( mesh_file ).c_str());
-        mesh->read_gmsh_from_stream(in);
+        mesh = mesh_full_constructor("{mesh_file=\"mesh/simplest_cube.msh\"}");
 
         Input::Type::Record rec_type = Input::Type::Record("Test","")
             .declare_key("scalar", ScalarField::get_input_type(), Input::Type::Default::obligatory(),"" )
             .declare_key("scalar_unit_conversion", ScalarField::get_input_type(), Input::Type::Default::obligatory(),"" )
             .declare_key("vector_fixed", VecFixField::get_input_type(), Input::Type::Default::obligatory(),"" )
             .declare_key("tensor_fixed", TensorField::get_input_type(), Input::Type::Default::obligatory(),"" )
+            .declare_key("vtk_scalar", ScalarField::get_input_type(), Input::Type::Default::obligatory(),"" )
+            .declare_key("vtk_vector", VecFixField::get_input_type(), Input::Type::Default::obligatory(),"" )
+            .declare_key("vtk_tensor", TensorField::get_input_type(), Input::Type::Default::obligatory(),"" )
             .close();
 
         Input::ReaderToStorage reader( input, rec_type, Input::FileFormat::format_JSON );
@@ -85,7 +100,7 @@ public:
 
     }
     virtual void TearDown() {
-
+    	delete mesh;
     }
 
     const FieldAlgoBaseInitData& init_data(std::string field_name) {
@@ -93,7 +108,7 @@ public:
     	return init_data;
     }
 
-    Mesh *mesh;
+    Mesh * mesh;
     Input::Record rec;
     Space<3>::Point point;
     double test_time[2];
@@ -118,8 +133,8 @@ TEST_F(FieldElementwiseTest, scalar) {
 
 TEST_F(FieldElementwiseTest, bc_scalar) {
     ScalarField field;
-    field.set_mesh(mesh,true);
     field.init_from_input(rec.val<Input::Record>("scalar"), init_data("scalar"));
+    field.set_mesh(mesh,true);
 
     for (unsigned int j=0; j<2; j++) {
     	field.set_time(test_time[j]);
@@ -147,8 +162,8 @@ TEST_F(FieldElementwiseTest, scalar_unit_conv) {
 
 TEST_F(FieldElementwiseTest, bc_scalar_unit_conv) {
     ScalarField field;
-    field.set_mesh(mesh,true);
     field.init_from_input(rec.val<Input::Record>("scalar_unit_conversion"), init_data("scalar_unit_conversion"));
+    field.set_mesh(mesh,true);
 
     for (unsigned int j=0; j<2; j++) {
     	field.set_time(test_time[j]);
@@ -233,6 +248,47 @@ TEST_F(FieldElementwiseTest, bc_tensor_fixed) {
 
 
 
+TEST_F(FieldElementwiseTest, vtk_scalar) {
+	ScalarField field;
+    field.init_from_input(rec.val<Input::Record>("vtk_scalar"), init_data("vtk_scalar"));
+    field.set_mesh(mesh, false);
+	field.set_time(0.0);
+
+	for(unsigned int i=0; i<mesh->element.size(); i++) {
+		EXPECT_DOUBLE_EQ( (i+1)*0.1 , field.value(point,mesh->element_accessor(i)) );
+	}
+}
+
+
+
+TEST_F(FieldElementwiseTest, vtk_vector) {
+	string expected_vals = "0.5 1 1.5";
+    VecFixField field;
+    field.init_from_input(rec.val<Input::Record>("vtk_vector"), init_data("vtk_vector"));
+    field.set_mesh(mesh, false);
+   	field.set_time(0.0);
+
+    for(unsigned int i=0; i < mesh->element.size(); i++) {
+    	EXPECT_TRUE( arma::min(arma::vec3(expected_vals) == field.value(point,mesh->element_accessor(i))) );
+    }
+}
+
+
+
+TEST_F(FieldElementwiseTest, vtk_tensor) {
+	string expected_vals = "1 4 7; 2 5 8; 3 6 9";
+    TensorField field;
+    field.init_from_input(rec.val<Input::Record>("vtk_tensor"), init_data("vtk_tensor"));
+    field.set_mesh(mesh,false);
+   	field.set_time(0.0);
+
+    for(unsigned int i=0; i < mesh->element.size(); i++) {
+    	arma::umat match = ( arma::mat33(expected_vals) == field.value(point,mesh->element_accessor(i)) );
+        EXPECT_TRUE( match.min() );
+    }
+}
+
+
 TEST_F(FieldElementwiseTest, scalar_enum) {
     EnumField field;
     field.set_mesh(mesh,false);
@@ -256,10 +312,10 @@ TEST_F(FieldElementwiseTest, bc_scalar_enum) {
     for (unsigned int j=0; j<2; j++) {
 		field.set_time(test_time[j]);
 
-		/*for(unsigned int i=0; i<6; i++) {
-			unsigned int val = i + j + ( i<4 ? 1 : 10 );
-			field.set_data_row(i, val );
-		}*/
+		//for(unsigned int i=0; i<6; i++) {
+		//	unsigned int val = i + j + ( i<4 ? 1 : 10 );
+		//	field.set_data_row(i, val );
+		//}
 		for(unsigned int i=0; i < 6; i++) {
 			EXPECT_EQ( (unsigned int)0, field.value(point,mesh->element_accessor(i,true)) );
 		}