+++ unit_tests/output/observe_test.cpp
@@ -16,6 +16,7 @@
 #include <mesh_constructor.hh>
 #include "io/observe.hh"
 #include "mesh/mesh.h"
+#include "io/msh_gmshreader.h"
 #include "input/reader_to_storage.hh"
 #include "input/accessors.hh"
 #include "system/sys_profiler.hh"
@@ -215,9 +216,7 @@ TEST(ObservePoint, find_observe_point) {
     armadillo_setup();
 
     FilePath mesh_file( string(UNIT_TESTS_SRC_DIR) + "/mesh/simplest_cube.msh", FilePath::input_file);
-    Mesh *mesh = mesh_constructor();
-    ifstream in(string(mesh_file).c_str());
-    mesh->read_gmsh_from_stream(in);
+    Mesh *mesh = mesh_full_constructor("{mesh_file=\"" + (string)mesh_file + "\"}");
 
     auto obs = TestObservePoint("0 -0.5 -0.5", 4, "ALL");
     obs.check(*mesh,"0.25 0.25 0.25", "0 -0.5 -0.5", 6);
@@ -243,14 +242,12 @@ TEST(Observe, all) {
         .get_root_interface<Input::Record>();
 
     FilePath mesh_file( string(UNIT_TESTS_SRC_DIR) + "/mesh/simplest_cube.msh", FilePath::input_file);
-    Mesh *mesh = mesh_constructor( "{ mesh_file=\"\", global_observe_search_radius=1.0 }" );
-    ifstream in(string(mesh_file).c_str());
-    mesh->read_gmsh_from_stream(in);
+    Mesh *mesh = mesh_full_constructor("{mesh_file=\"" + (string)mesh_file + "\", global_observe_search_radius=1.0 }");
 
     {
-    TestObserve obs(*mesh, in_rec.val<Input::Array>("observe_points"));
-    obs.check_points_input();
-    obs.check_observe_points();
+    std::shared_ptr<TestObserve> obs = std::make_shared<TestObserve>(*mesh, in_rec.val<Input::Array>("observe_points"));
+    obs->check_points_input();
+    obs->check_observe_points();
 
     // read fiels
     TimeGovernor tg(0.0, 1.0);
@@ -258,19 +255,19 @@ TEST(Observe, all) {
     field_set.set_input_list( in_rec.val<Input::Array>("input_fields") );
     field_set.set_time(tg.step(), LimitSide::right);
 
-    obs.compute_field_values(field_set.scalar_field);
-    obs.compute_field_values(field_set.enum_field);
-    obs.compute_field_values(field_set.vector_field);
-    obs.compute_field_values(field_set.tensor_field);
-    obs.output_time_frame( tg.t() );
+    field_set.scalar_field.observe_output(obs);
+    field_set.enum_field.observe_output(obs);
+    field_set.vector_field.observe_output(obs);
+    field_set.tensor_field.observe_output(obs);
+    obs->output_time_frame( tg.t() );
 
     tg.next_time();
     field_set.set_time( tg.step(), LimitSide::right);
-    obs.compute_field_values(field_set.scalar_field);
-    obs.compute_field_values(field_set.enum_field);
-    obs.compute_field_values(field_set.vector_field);
-    obs.compute_field_values(field_set.tensor_field);
-    obs.output_time_frame( tg.t() );
+    field_set.scalar_field.observe_output(obs);
+    field_set.enum_field.observe_output(obs);
+    field_set.vector_field.observe_output(obs);
+    field_set.tensor_field.observe_output(obs);
+    obs->output_time_frame( tg.t() );
     }
     // closed observe file 'test_eq_observe.yaml'
     // check results