+++ unit_tests/output/output_vtk_test.cpp
@@ -17,9 +17,11 @@
 #include "io/output_vtk.hh"
 #include "io/output_mesh.hh"
 #include "mesh/mesh.h"
+#include "io/msh_gmshreader.h"
 #include "input/reader_to_storage.hh"
 #include "system/sys_profiler.hh"
 #include "system/logger_options.hh"
+#include "fields/field.hh"
 
 FLOW123D_FORCE_LINK_IN_PARENT(field_constant)
 
@@ -36,7 +38,7 @@ format: !vtk
 )YAML";
 
 
-class TestOutputVTK : public testing::Test, public OutputVTK {
+class TestOutputVTK : public OutputVTK, public std::enable_shared_from_this<OutputVTK> {
 public:
     TestOutputVTK()
     : OutputVTK()
@@ -45,9 +47,7 @@ public:
         LoggerOptions::get_instance().set_log_file("");
 
         FilePath mesh_file( string(UNIT_TESTS_SRC_DIR) + "/fields/simplest_cube_3d.msh", FilePath::input_file);
-        this->_mesh = mesh_constructor();
-        ifstream in(string(mesh_file).c_str());
-        this->_mesh->read_gmsh_from_stream(in);
+        this->_mesh = mesh_full_constructor("{mesh_file=\"" + (string)mesh_file + "\"}");
 
         component_names = { "comp_0", "comp_1", "comp_2" };
     }
@@ -91,7 +91,7 @@ public:
         this->output_mesh_discont_ = std::make_shared<OutputMeshDiscontinuous>( *(this->_mesh) );
         this->output_mesh_discont_->create_mesh(this->output_mesh_);
 
-		this->compute_field_data(ELEM_DATA, field);
+		field.compute_field_data(ELEM_DATA, shared_from_this());
 	}
 
 	// check result
@@ -110,33 +110,54 @@ public:
 	    EXPECT_EQ(str_vtk_file_ref.str(), str_vtk_file.str());
 	}
 
+	void set_current_step(int step) {
+		this->current_step = step;
+	}
+
+	std::string base_filename() {
+		return string(this->_base_filename);
+	}
+
+	std::string main_output_basename() {
+		return this->main_output_basename_;
+	}
+
+	std::string main_output_dir() {
+		return this->main_output_dir_;
+	}
+
 	std::vector<string> component_names;
 };
 
 
-TEST_F(TestOutputVTK, write_data_ascii) {
-	//( const_cast<Input::Type::Instance &>(FieldAlgorithmBase<3, FieldValue<3>::VectorFixed >::get_input_type_instance()) ).finish(Input::Type::FinishType::root_of_generic);
+TEST(TestOutputVTK, write_data_ascii) {
+	std::shared_ptr<TestOutputVTK> output_vtk = std::make_shared<TestOutputVTK>();
 
-	this->init_mesh(test_output_time_ascii);
-    this->current_step=1;
-    this->write_data();
-    EXPECT_EQ("./test1.pvd", string(this->_base_filename));
-    EXPECT_EQ("test1", this->main_output_basename_);
-    EXPECT_EQ(".", this->main_output_dir_);
+	output_vtk->init_mesh(test_output_time_ascii);
+	output_vtk->set_current_step(1);
+	output_vtk->set_field_data< Field<3,FieldValue<0>::Scalar> > ("scalar_field", "0.5");
+	output_vtk->set_field_data< Field<3,FieldValue<3>::VectorFixed> > ("vector_field", "[0.5, 1.0, 1.5]");
+	output_vtk->set_field_data< Field<3,FieldValue<3>::TensorFixed> > ("tensor_field", "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
+	output_vtk->write_data();
+    EXPECT_EQ("./test1.pvd", output_vtk->base_filename());
+    EXPECT_EQ("test1", output_vtk->main_output_basename());
+    EXPECT_EQ(".", output_vtk->main_output_dir());
 
-    check_result_file("test1/test1-000001.vtu", "test_output_vtk_ascii_ref.vtu");
+    output_vtk->check_result_file("test1/test1-000001.vtu", "test_output_vtk_ascii_ref.vtu");
 }
 
 
-TEST_F(TestOutputVTK, write_data_binary) {
-	this->init_mesh(test_output_time_binary);
-    this->current_step=0;
-    set_field_data< Field<3,FieldValue<0>::Scalar> > ("scalar_field", "0.5");
-    set_field_data< Field<3,FieldValue<3>::VectorFixed> > ("vector_field", "[0.5, 1.0, 1.5]");
-    set_field_data< Field<3,FieldValue<3>::TensorFixed> > ("tensor_field", "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
-    this->write_data();
+TEST(TestOutputVTK, write_data_binary) {
+	std::shared_ptr<TestOutputVTK> output_vtk = std::make_shared<TestOutputVTK>();
 
-    check_result_file("test1/test1-000000.vtu", "test_output_vtk_binary_ref.vtu");
+	output_vtk->init_mesh(test_output_time_binary);
+    output_vtk->set_current_step(0);
+    output_vtk->set_field_data< Field<3,FieldValue<0>::Scalar> > ("scalar_field", "0.5");
+    output_vtk->set_field_data< Field<3,FieldValue<3>::VectorFixed> > ("vector_field", "[0.5, 1.0, 1.5]");
+    output_vtk->set_field_data< Field<3,FieldValue<3>::TensorFixed> > ("tensor_field", "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
+    output_vtk->write_data();
+
+    output_vtk->check_result_file("test1/test1-000000.vtu", "test_output_vtk_binary_ref.vtu");
 }
 
 #ifdef FLOW123D_HAVE_ZLIB
@@ -147,15 +168,17 @@ format: !vtk
   variant: binary_zlib
 )YAML";
 
-TEST_F(TestOutputVTK, write_data_compressed) {
-	this->init_mesh(test_output_time_compressed);
-    this->current_step=0;
-    set_field_data< Field<3,FieldValue<0>::Scalar> > ("scalar_field", "0.5");
-    set_field_data< Field<3,FieldValue<3>::VectorFixed> > ("vector_field", "[0.5, 1.0, 1.5]");
-    set_field_data< Field<3,FieldValue<3>::TensorFixed> > ("tensor_field", "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
-    this->write_data();
+TEST(TestOutputVTK, write_data_compressed) {
+	std::shared_ptr<TestOutputVTK> output_vtk = std::make_shared<TestOutputVTK>();
 
-    check_result_file("test1/test1-000000.vtu", "test_output_vtk_zlib_ref.vtu");
+	output_vtk->init_mesh(test_output_time_compressed);
+    output_vtk->set_current_step(0);
+    output_vtk->set_field_data< Field<3,FieldValue<0>::Scalar> > ("scalar_field", "0.5");
+    output_vtk->set_field_data< Field<3,FieldValue<3>::VectorFixed> > ("vector_field", "[0.5, 1.0, 1.5]");
+    output_vtk->set_field_data< Field<3,FieldValue<3>::TensorFixed> > ("tensor_field", "[[1, 2, 3], [4, 5, 6], [7, 8, 9]]");
+    output_vtk->write_data();
+
+    output_vtk->check_result_file("test1/test1-000000.vtu", "test_output_vtk_zlib_ref.vtu");
 }
 
 #endif // FLOW123D_HAVE_ZLIB